refactor: move location of `KindeContext`
